Support the usage of jTDS driver.
If users want to use Microsoft SQLServerDriver, they could specify the driver's jar file path as 'driver_path'
field or put the jar file on the classpath of the plugin's classloader. The plugin can read the jar file and
use it. Otherwise, jTDS driver can be used for the plugin without any configuration.